Move ThumbPrint into VMwareWebService
