+++ lib/gems/pending/VMwareWebService/esx_thumb_print.rb
@@ -1,4 +1,4 @@
-require 'util/thumb_print'
+require 'VMwareWebService/thumb_print'
 
 class ESXThumbPrint < ThumbPrint
   attr_reader :user, :password
