+++ lib/gems/pending/VMwareWebService/thumb_print.rb
